results-code/requirements.txt
# using python3.12 
# jpype is optional dependency; tabula-py throws non-fatal error without it

tabula-py
jpype1
pandas

